+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## master
 
+### Fixes
+
+- `[expect]` `toEqual` no longer tries to compare non-enumerable symbolic properties, to be consistent with non-symbolic properties. ([#6398](https://github.com/facebook/jest/pull/6398))
+
 ## 23.1.0
 
 ### Features